Implemented system level testing and modify the component testing
